Minor code cleanup
